finish project

+++ frontend/src/components/WorkoutDetials.jsx
@@ -1,13 +1,23 @@
-import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
-import {RiDeleteBinLine} from 'react-icons/ri';
-import  formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
+import {MdDelete} from 'react-icons/md';
+// date fns
+import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if (!user) {
+      return
+    }
+
     const response = await fetch('/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
 
@@ -20,9 +30,9 @@ const WorkoutDetails = ({ workout }) => {
     <div className="workout-details">
       <h4>{workout.title}</h4>
       <p><strong>Load (kg): </strong>{workout.load}</p>
-      <p><strong>Number of reps: </strong>{workout.reps}</p>
-      <p>{formatDistanceToNow(new Date (workout.createdAt) , {addSuffix : true})}</p>
-      <span onClick={handleClick}><RiDeleteBinLine/></span>
+      <p><strong>Reps: </strong>{workout.reps}</p>
+      <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
+      <span className="material-symbols-outlined" onClick={handleClick}><MdDelete /></span>
     </div>
   )
 }
